requirements.txt
PyQt6
pyqtgraph
pyModbusTCP
numpy
labjack-ljm
